some updates
Removed Horse Stats Vanilla, updated a bunch of other mods

unfortunately we're still missing a bunch of mods for 1.18.2. here are my thoughts on each mod that has yet to be updated.

also i'm not listing the versions of the mods that were updated. figure it out for yourself fucko. ANYWAYS:

BetterEnd: i have been considering removing it because i don't want to go too overboard with adding new materials and this mod actually does add new materials that end up being stronger than netherite. i want the pack to keep a sort-of vanilla feel and adding new materials pushes it further away from that. however the mod does add some cool shit biome-wise.

Botania: to be completely honest I've never actually played around with the mod while testing the pack out. however it is one of those older mods that's decently well-known and the fact that i can have one of those older mods in the pack is really cool.

Gate of Babylon: I REALLY want to keep this mod because so far it's been the only mod i could find that adds new types of weapons. unfortunately it hasn't been updated since february (before 1.18.2 came out) which is painful.

Rat's Mischief: honestly we could do without it, i'd just like to have it in because ha ha funny rat

Spectrum: i really want to keep this mod in because it adds so much to the game yet keeps the game feel a lot like vanilla, as you progressively unlock the ability to see certain blocks. the only issue i really have with it is the geodes spawning above ground, which i can easily fix through the configs.

Stoneholm: underground villages. a relatively minor addition, but one that is really nice to have nonetheless.

YUNG's Better Strongholds: I'd rather have this in the pack for fairly obvious reasons. It genuinely makes strongholds better and more interesting than the current vanilla design which has been around for over a decade now.

anyways yeah